Create Project5.sql
In the project 5, I have used 2 CTE subclause. One to calculate the average salary of the all the employees of the organisation and the other subclause is to calculate the maximum salary of each female working in the company. Then i have compared both of them according to the demand of the question.